fix: avoid error notification when user cancels screenshot
Displaying an error notification when user decides to back out from
taking a screenshot is not a proper behavior. The user can cancel 
the screenshot by either pressing ESC or clicking on the "x" button.

I am not sure if there's any way to actually discover the fact that 
the user wanted to cancel the screenshot. I just see that the
error gets set when the user cancels. I'm not sure what other
kinds of errors might occur.